Fix libriscv compilation on Apple platforms
- Fix undeclared g_epoll_instances error in system_calls.cpp
- Add platform compatibility layer (platform_compat.hpp)
- Add hardware abstraction layer interface (platform_hal.hpp)
- Move Apple epoll emulation declarations to proper scope
- Fix timespec namespace conflicts with ::timespec qualifier
- Add proper cleanup of epoll instances in syscall_close
- Include required headers (time.h, sys/event.h, Security.h)
- Implement complete epoll emulation using kqueue for macOS/iOS
- Fix kevent function signature for timeout parameter

Resolves compilation errors on both macOS and iOS platforms.
Tested with successful builds on both platforms.

+++ modules/sandbox/src/sandbox_syscalls.cpp
@@ -270,7 +270,7 @@ APICALL(api_vcreate) {
 					const unsigned size = method;
 					// Copy array of Variants from guest memory.
 					const GuestVariant *gvec = machine.memory.memarray<const GuestVariant>(gdata, size);
-					for (int i = 0; i < size; i++) {
+					for (unsigned i = 0; i < size; i++) {
 						a.push_back(gvec[i].toVariant(emu));
 					}
 				} else {
@@ -1944,7 +1944,7 @@ APICALL(api_sandbox_add) {
 			if (Ref<ELFScript> program = emu.get_program(); program.is_valid()) {
 				Dictionary func = Sandbox::create_public_api_function(name, address, description, return_type, arguments);
 				if (func.size() > 0) {
-					if (program->functions.size() >= Sandbox::MAX_PUBLIC_FUNCTIONS) {
+					if (static_cast<int>(program->functions.size()) >= Sandbox::MAX_PUBLIC_FUNCTIONS) {
 						ERR_PRINT("Too many public functions in the Sandbox program");
 						throw std::runtime_error("Too many public functions in the Sandbox program");
 					}
